create game logic header
