+++ snake/src/gameLogic.c
@@ -1,28 +1,10 @@
-
-#include <stdbool.h>
+#include "gameLogic.h"
 #include <stdint.h>
 #include <stdlib.h>
 
-#define BOARD_WIDHT 16
-#define BOARD_HEIGHT 10
-
-int initGame();
-int initSnake();
-int initPuff();
-int tickGame(enum Direction);
-void movePuff();
 bool compareCoords(Pokepuff *, SnakeHead *);
 bool compareCoords(Pokepuff *, SnakeBody *);
 bool compareCoords(SnakeHead *, SnakeBody *);
-bool isSnekDead();
-void cleanGame();
-
-struct TileLocation {
-    int x, y;
-};
-
-// its useful for opposite directions to be negatives of one another
-enum Direction { UP = 1, DOWN = -1, LEFT = 2, RIGHT = -2 };
 
 typedef struct {
     int x, y;
